+++ src/main/java/domain/order/Category.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.order;
 
 public enum Category {
     CHICKEN("치킨"),
