+++ src/main/java/domain/order/MenuRepository.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.order;
 
 import java.util.ArrayList;
 import java.util.Collections;
