+++ src/main/java/domain/order/Order.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.order;
 
 public class Order {
 	private final Menu menu;
