+++ src/main/java/domain/order/Orders.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.order;
 
 import java.util.List;
