+++ src/test/java/domain/order/AmountTest.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.order;
 
+import domain.order.Amount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
