+++ src/test/java/domain/order/OrderTest.java
@@ -1,5 +1,8 @@
-package domain;
+package domain.order;
 
+import domain.order.Amount;
+import domain.order.MenuRepository;
+import domain.order.Order;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
