+++ src/test/java/domain/order/OrdersTest.java
@@ -1,5 +1,9 @@
-package domain;
+package domain.order;
 
+import domain.order.Amount;
+import domain.order.MenuRepository;
+import domain.order.Order;
+import domain.order.Orders;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
